dashApp/requirements.txt
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
gunicorn==20.0.4
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
pyzmq==20.0.0
requests==2.25.1
retrying==1.3.3
six==1.15.0
urllib3==1.26.5
Werkzeug==2.0.1
